requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Machine Learning
lightgbm>=4.0.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Image/Video processing
imageio>=2.31.0
imageio-ffmpeg>=0.4.9
Pillow>=10.0.0

# Progress bars
tqdm>=4.65.0

# Optional: SHAP for model interpretation
shap>=0.42.0
